poor time zones are getting eaten
