+++ src/components/Tasks/TaskDateInput.js
@@ -1,17 +1,11 @@
+import { minMaxDate } from "utils/DateUtils"
 
 
 export const TaskDateInput = ({info, onDateChange})=> {
     const handleDateChanged = (event) => {
       onDateChange(event.target.value)
     }
-  
-      const today = new Date()
-      const minDate = new Date();
-      const maxDate = new Date()
-  
-      minDate.setFullYear(today.getFullYear() - 1);
-      maxDate.setFullYear(today.getFullYear() + 5)
-  
+    const [minDate, maxDate] = minMaxDate()
       return (
           <form>
             <label>
